Removed Mp3agic in favor of eye3D, MySQL tables refactoring, fixed bug casting long to int, fixed multiple datasources creation making Guice injection thread safe

+++ filemanager/src/main/java/models/Model.java
@@ -33,13 +33,13 @@ public abstract class Model<T> {
 
     protected void positiveNumberValidator(String fieldName, Short num) {
         if (num != null) {
-            this.positiveNumberValidator(fieldName, num.intValue());
+            this.positiveNumberValidator(fieldName, num);
         }
     }
 
     protected void positiveNumberValidator(String fieldName, Long num) {
         if (num != null) {
-            this.positiveNumberValidator(fieldName, num.intValue());
+            this.positiveNumberValidator(fieldName, num);
         }
     }
